Created new temp file for implementation of sprites and groups.
Created a new Player2 and Buff class for testing.
Collision between Player2 and Buff works as intended.

Currently Player2 is an "overlay" over original Player class. Please look through the changes and possibly replace Player with Player2.

Enemy should be changed to sprite and be added to a group. Collosion between enemies should be easier with Sprite and Group method.